Transferred the tests to use the components instead
