--- Plugins/EnhancedSaveGame/Source/EnhancedSaveGameTests/Private/FunctionalTesting/SaveSystemFunctionalTest.cpp
@@ -1,8 +1,0 @@
-﻿// "Unreal Pokémon" created by Retro & Chill.
-
-#include "FunctionalTesting/SaveSystemFunctionalTest.h"
-#include "Saving/SaveGameManager/ScopedInMemorySaveHandle.h"
-
-void ASaveSystemFunctionalTest::AddScopedTestObjects() {
-    AddScopedData<EnhancedSaveGame::FScopedInMemorySaveHandle>();
-}
