+++ Plugins/EnhancedSaveGame/Source/EnhancedSaveGameTests/Public/FunctionalTesting/SaveSystemTestComponent.h
@@ -3,11 +3,18 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "ScopedTestComponent.h"
+#include "FunctionalTesting/ScopedTestComponent.h"
 #include "SaveSystemTestComponent.generated.h"
 
 
-UCLASS(ClassGroup=(Custom), meta=(BlueprintSpawnableComponent))
+/**
+ * @class USaveSystemTestComponent
+ * @brief A component class for handling enhanced save game tests.
+ *
+ * Inherits from UScopedTestComponent to integrate enhanced save game functionality
+ * within the Unreal Engine testing framework.
+ */
+UCLASS(ClassGroup=(FunctionalTest), meta=(BlueprintSpawnableComponent))
 class ENHANCEDSAVEGAMETESTS_API USaveSystemTestComponent : public UScopedTestComponent {
     GENERATED_BODY()
